OK, no have data showing up in a struct form.
This seems to be getting closer to the answer. The data is in byte form.
It requires the data to be changed from bytes to a struct. Now, to
access all of the different data points as they come in. I feel like
this should be a sort of ticker, with continuous incoming data. Given
that the data is pretty simple, I just have to be able to display it in
a proper manner. Thinking of a sort of table, but haven't found the
right one yet. Will have to do some googling.